Refactor to allow for multiple meshes
Refactored cube_mesh into a generic mesh module that can handle an array of meshes that can be drawn to the screen.

+++ src/scene_instructions.c
@@ -34,7 +34,7 @@ bool instructions_handle_keyboard(SDL_Event event) {
 	return false;
 }
 
-void instructions_update(uint64_t delta_time) {
+void instructions_update(double delta_time) {
 	
 }
